refactor: 意味のないpackage分けを削除した
