--- lib/errors/doc.go
@@ -1,2 +1,0 @@
-// Package errors はゲーム固有のエラー型とユーティリティを提供する。
-package errors
