+++ lib/raw/errors.go
@@ -1,4 +1,4 @@
-package errors
+package raw
 
 import "fmt"
